+++ plugins/commands/reload/command.rb
@@ -33,8 +33,24 @@ module VagrantPlugins
         validate_provisioner_flags!(options)
 
         @logger.debug("'reload' each target VM...")
+        machines = []
         with_target_vms(argv) do |machine|
+          machines << machine
           machine.action(:reload, options)
+        end
+
+        # Output the post-up messages that we have, if any
+        machines.each do |m|
+          next if !m.config.vm.post_up_message
+          next if m.config.vm.post_up_message == ""
+
+          # Add a newline to separate things.
+          @env.ui.info("", prefix: false)
+
+          m.ui.success(I18n.t(
+            "vagrant.post_up_message",
+            name: m.name.to_s,
+            message: m.config.vm.post_up_message))
         end
 
         # Success, exit status 0